+++ apps/swirl-docs/src/components/CodePreview/CodePreview.tsx
@@ -1,153 +1,98 @@
-import {
-  SwirlIconCheckStrong,
-  SwirlIconCopy,
-  SwirlIconExpandLess,
-  SwirlIconExpandMore,
-} from "@getflip/swirl-components-react";
-import { ReactNode, useState } from "react";
+import { useEffect, useState } from "react";
 import { CodeSandboxButton } from "./CodeSandboxButton";
-import { CopyToClipboard } from "react-copy-to-clipboard";
 import classNames from "classnames";
 import NoSsr from "../Layout/NoSsr";
-import Highlight, { defaultProps } from "prism-react-renderer";
-import darkTheme from "prism-react-renderer/themes/vsDark";
-import lightTheme from "prism-react-renderer/themes/nightOwlLight";
-import { CodeExample } from "./types";
+import { CodePreview } from "./types";
 import { NpmPackageLink } from "./NpmPackageLink";
 import CodePreviewContext from "./CodePreviewContext";
 import { APIEndpointHeader } from "./CodePreviewRequestString";
-
-interface CodePreviewProps {
-  codeExample: CodeExample;
-  isHttpResponse?: boolean;
-  children?: ReactNode | ReactNode[];
-}
+import { CodePreviewHeader } from "./CodePreviewHeader";
+import { CodePreviewHighlight } from "./CodePreviewHighlight";
+import { CodePreviewExpandButton } from "./CodePreviewExpandButton";
 
 /**
  * Let's you easily render syntax highlighted code.
- * Children of the Component will be displayed in the Header.
  */
 export function CodePreview({
-  children,
   codeExample,
-  isHttpResponse,
-}: CodePreviewProps) {
+  isLightTheme,
+  hasCopyButton,
+  className,
+  PreviewIndicator,
+  MainHeaderContent,
+  ActionItems,
+}: CodePreview) {
   const [isExpanded, setIsExpanded] = useState<boolean>(false);
-  const [isCopied, setIsCopied] = useState<boolean>(false);
+  const [selectId, handleSelect] = useState<
+    CodePreview["codeExample"]["selectedId"]
+  >(codeExample.selectedId);
+
+  const [codePreviewCodeExample, setCodeExample] = useState<
+    CodePreview["codeExample"]
+  >({
+    code: codeExample.code,
+    isLongCode: codeExample.isLongCode,
+    selectedId: selectId,
+  });
+
+  useEffect(() => {
+    // TODO: Heads up: not the best solution for now. A global reducer would be better for the long run.
+    if (codeExample && !selectId) {
+      setCodeExample({
+        ...codeExample,
+        code: codeExample.code,
+        isLongCode: codeExample.code.split("\n").length > 7,
+      });
+    }
+    if (codeExample.selectOptions && selectId) {
+      setCodeExample({
+        ...codeExample,
+        selectedId: selectId,
+        code: codeExample.selectOptions[selectId],
+        isLongCode: codeExample.selectOptions[selectId].split("\n").length > 7,
+      });
+    }
+  }, [selectId, codeExample]);
 
   return (
     <NoSsr>
-      <CodePreviewContext.Provider value={codeExample}>
+      <CodePreviewContext.Provider
+        value={{
+          isLightTheme,
+          codeExample: codePreviewCodeExample,
+          hasCopyButton,
+          className,
+          isExpanded,
+          PreviewIndicator,
+          MainHeaderContent,
+          ActionItems,
+          handleSelect,
+        }}
+      >
         <div
           className={classNames(
-            "relative w-auto  rounded-lg mb-10 overflow-auto",
-            {
-              "bg-[#24292E]": !isHttpResponse,
-              "bg-surface-raised-default": isHttpResponse,
-            },
+            className,
+            "box-border relative max-w-full w-full rounded-xl",
+            "p-2",
             {
               "md:h-[240px] md:max-h-[240px] overflow-hidden":
-                !isExpanded && !isHttpResponse,
+                !isExpanded && !isLightTheme,
               "min-h-[240px]": isExpanded,
-              "h-full": isHttpResponse,
+              "h-full": isLightTheme,
+            },
+            {
+              "bg-[#24292E]": !isLightTheme,
+              "bg-surface-raised-default": isLightTheme,
             }
           )}
         >
-          <div
-            className={classNames(
-              "hidden md:flex items-center justify-between h-12 m-2 p-4 rounded-lg",
-              {
-                "bg-[#21201E]": !isHttpResponse,
-                "bg-surface-overlay-default": isHttpResponse,
-              }
-            )}
-          >
-            <div className="flex items-center justify-between">{children}</div>
-            {!isHttpResponse && (
-              <CopyToClipboard
-                text={codeExample.code}
-                onCopy={() => {
-                  setIsCopied(true);
-                  setTimeout(() => {
-                    setIsCopied(false);
-                  }, 2000);
-                }}
-              >
-                <button
-                  aria-label="Copy code to clipboard"
-                  type="button"
-                  className="flex justify-center items-center text-[#F2F2F2] text-base font-medium"
-                >
-                  {isCopied ? "Code copied!" : "Copy Code"}
-                  {isCopied ? (
-                    <SwirlIconCheckStrong size={16} className="ml-1" />
-                  ) : (
-                    <SwirlIconCopy size={16} className="ml-1" />
-                  )}
-                </button>
-              </CopyToClipboard>
-            )}
-          </div>
-
-          <Highlight
-            {...defaultProps}
-            theme={isHttpResponse ? lightTheme : darkTheme}
-            code={codeExample.code}
-            language={codeExample.language ? codeExample.language : "tsx"}
-          >
-            {({ tokens, getLineProps, getTokenProps }) => (
-              <pre
-                className={classNames(
-                  "cursor-text overflow-auto pt-space-16 md:pt-space-8 px-space-24 ",
-                  "md:pb-16",
-                  {
-                    "pb-16": isExpanded,
-                    "pb-space-16": !isExpanded || isHttpResponse,
-                    "md:pb-space-16": isHttpResponse,
-                  }
-                )}
-              >
-                <code>
-                  {tokens.map((line, i) => (
-                    <div key={i} {...getLineProps({ line, key: i })}>
-                      {line.map((token, key) => (
-                        <span key={i} {...getTokenProps({ token, key })} />
-                      ))}
-                    </div>
-                  ))}
-                </code>
-              </pre>
-            )}
-          </Highlight>
-
-          {codeExample.isLongCode && !isHttpResponse && (
-            <div
-              className={classNames(
-                "absolute bottom-0 flex justify-center items-center w-full h-12",
-                {
-                  "bg-[#24292E]": !isHttpResponse,
-                  "bg-surface-raised-default": isHttpResponse,
-                }
-              )}
-            >
-              <button
-                onClick={() => setIsExpanded(!isExpanded)}
-                className={classNames(
-                  "bottom-2 left-auto right-auto flex justify-center items-center text-[#F2F2F2] text-base font-medium",
-                  {
-                    "text-[#F2F2F2]": !isHttpResponse,
-                    "text-[#24292E]": isHttpResponse,
-                  }
-                )}
-              >
-                {isExpanded ? "Collapse" : "Expand"}
-                {isExpanded ? (
-                  <SwirlIconExpandLess className="ml-2" size={24} />
-                ) : (
-                  <SwirlIconExpandMore className="ml-2" size={24} />
-                )}
-              </button>
-            </div>
+          <CodePreviewHeader />
+          <CodePreviewHighlight />
+          {codePreviewCodeExample.isLongCode && !isLightTheme && (
+            <CodePreviewExpandButton
+              isExpanded={isExpanded}
+              setIsExpanded={setIsExpanded}
+            />
           )}
         </div>
       </CodePreviewContext.Provider>
@@ -157,4 +102,4 @@ export function CodePreview({
 
 CodePreview.CodeSandboxButton = CodeSandboxButton;
 CodePreview.NpmPackageLink = NpmPackageLink;
-CodePreview.Request = APIEndpointHeader;
+CodePreview.EndpointHeader = APIEndpointHeader;