+++ apps/swirl-docs/src/components/Layout/DocumentationLayout.tsx
@@ -30,11 +30,11 @@ export function DocumentationLayout({
     <DocumentationLayoutContext.Provider value={data}>
       <div className="grid grid-cols-1 lg:grid-cols-documentation-layout h-full overflow-hidden">
         {data.navigationLinks && <CategoryNav />}
-        <div className="w-full h-full overflow-auto scroll-p-4">
+        <div className="w-full h-full overflow-auto scroll-p-4 flex flex-col items-center">
           <main
             id="main"
             className={classNames(
-              "grid grid-cols-1 justify-center",
+              "grid grid-cols-1 justify-center max-w-[77.5rem]",
               "my-0 mx-auto mt-14 mb-4 md:mb-0 px-4 lg:px-0",
               {
                 "md:grid-cols-1 md:mx-10": disableToc,