+++ apps/swirl-docs/src/pages/_document.tsx
@@ -1,5 +1,4 @@
 import Document, { Html, Head, Main, NextScript } from "next/document";
-import { GTM_ID, isProd } from "src/lib/gtm";
 
 class MyDocument extends Document {
   render() {
@@ -24,16 +23,6 @@ class MyDocument extends Document {
           <link rel="mask-icon" href="/safari-pinned-tab.svg" color="#145af5" />
         </Head>
         <body>
-          {isProd && (
-            <noscript>
-              <iframe
-                src={`https://www.googletagmanager.com/ns.html?id=${GTM_ID}`}
-                height="0"
-                width="0"
-                style={{ display: "none", visibility: "hidden" }}
-              />
-            </noscript>
-          )}
           <Main />
           <div id="portal" />
           <NextScript />