+++ apps/swirl-docs/src/pages/api/tokens/v0/utils.ts
@@ -1,4 +1,3 @@
-import { isProd } from "@swirl/lib/gtm";
 import fs from "fs";
 import path from "path";
 
@@ -17,6 +16,8 @@ export const lessDark = loadFile("/less/styles.dark.less");
 
 export const scssLight = loadFile("/scss/styles.light.scss");
 export const scssDark = loadFile("/scss/styles.dark.scss");
+
+export const isProd = process.env.NODE_ENV === "production";
 
 function loadFile(filePath: string) {
   let nodeModuleRoot = "node_modules/@getflip/swirl-tokens/dist";